Solution to Check if Two Strings are Anagrams

+++ Check if Two Strings are Anagrams@@676c039d80e1714600f6764a/code.c
@@ -13,7 +13,7 @@ int main() {
             }
         }
     }
-    printf("%d",(int)strlen(a));
+    printf("%lu",strlen(a));
 
     // if(strlen(a)==strlen(b) == e)  {
     //     printf("Yes");
